Kezdi.jl Documentation

Getting started

Installation

Kezdi.jl is currently in beta. To install the package, run the following command in Julia's REPL:

using Pkg; Pkg.add(url="https://github.com/codedthinking/Kezdi.jl#0.4-beta")

Example

using Kezdi
@use "test/test.dta"

@generate log_y = log(y)
@collapse mean_x = mean(x) mean_log_y = mean(log_y), by(s)
@regress mean_log_y mean_x, robust

<script async data-uid="62d7ebb237" src="https://relentless-producer-1210.ck.page/62d7ebb237/index.js"></script>

Benefits of using Kezdi.jl

Speed

Command	Stata	Julia 1st run	Julia 2nd run	Speedup

@egen	4.90s	1.60s	0.41s	10x

@collapse	0.92s	0.18s	0.13s	8x

@tabulate	2.14s	0.46s	0.10s	20x

@summarize	10.40s	0.58s	0.37s	28x

@regress	0.89s	1.93s	0.16s	6x

Commands

Filtering columns and rows

Kezdi.@keep — Macro

@keep y1 y2 ... [@if condition]

Keep only the variables y1, y2, etc. in df. If condition is provided, only the rows for which the condition is true are kept.

Kezdi.@drop — Macro

@drop y1 y2 ...

or @drop if condition]

Drop the variables y1, y2, etc. from df. If condition is provided, the rows for which the condition is true are dropped.

Modifying columns

Kezdi.@generate — Macro

@generate y = expr [@if condition]

Create a new variable y in df by evaluating expr. If condition is provided, the operation is executed only on rows for which the condition is true. When the condition is false, the variable will be missing.

Kezdi.@replace — Macro

@replace y = expr [@if condition]

Replace the values of y in df with the result of evaluating expr. If condition is provided, the operation is executed only on rows for which the condition is true. When the condition is false, the variable will be left unchanged.

Kezdi.@egen — Macro

@egen y1 = expr1 y2 = expr2 ... [@if condition], [by(group1, group2, ...)]

Generate new variables in df by evaluating expressions expr1, expr2, etc. If condition is provided, the operation is executed only on rows for which the condition is true. When the condition is false, the variables will be missing. If by is provided, the operation is executed by group.

Grouping data

Kezdi.@collapse — Macro

@collapse y1 = expr1 y2 = expr2 ... [@if condition], [by(group1, group2, ...)]

Collapse df by evaluating expressions expr1, expr2, etc. If condition is provided, the operation is executed only on rows for which the condition is true. If by is provided, the operation is executed by group.

Kezdi.@tabulate — Macro

@tabulate y1 y2 ... [@if condition]

Create a frequency table for the variables y1, y2, etc. in df. If condition is provided, the operation is executed only on rows for which the condition is true.

Kezdi.@summarize — Macro

@summarize y [@if condition]

Summarize the variable y in df. If condition is provided, the operation is executed only on rows for which the condition is true.

Kezdi.@regress — Macro

@regress y x1 x2 ... [@if condition], [robust] [cluster(var1, var2, ...)]

Estimate a regression model in df with dependent variable y and independent variables x1, x2, etc. If condition is provided, the operation is executed only on rows for which the condition is true. If robust is provided, robust standard errors are calculated. If cluster is provided, clustered standard errors are calculated.

With Module

Kezdi.With.@with — Macro

@with(expr, exprs...)

Rewrites a series of expressions into a with, where the result of one expression is inserted into the next expression following certain rules.

Rule 1

Any expr that is a begin ... end block is flattened. For example, these two pseudocodes are equivalent:

@with a b c d e f

@with a begin
    b
    c
    d
end e f

Rule 2

Any expression but the first (in the flattened representation) will have the preceding result inserted as its first argument, unless at least one underscore _ is present. In that case, all underscores will be replaced with the preceding result.

If the expression is a symbol, the symbol is treated equivalently to a function call.

For example, the following code block

@with begin
    x
    f()
    @g()
    h
    @i
    j(123, _)
    k(_, 123, _)
end

is equivalent to

begin
    local temp1 = f(x)
    local temp2 = @g(temp1)
    local temp3 = h(temp2)
    local temp4 = @i(temp3)
    local temp5 = j(123, temp4)
    local temp6 = k(temp5, 123, temp5)
end

Rule 3

An expression that begins with @aside does not pass its result on to the following expression. Instead, the result of the previous expression will be passed on. This is meant for inspecting the state of the with. The expression within @aside will not get the previous result auto-inserted, you can use underscores to reference it.

@with begin
    [1, 2, 3]
    filter(isodd, _)
    @aside @info "There are $(length(_)) elements after filtering"
    sum
end

Rule 4

It is allowed to start an expression with a variable assignment. In this case, the usual insertion rules apply to the right-hand side of that assignment. This can be used to store intermediate results.

@with begin
    [1, 2, 3]
    filtered = filter(isodd, _)
    sum
end

filtered == [1, 3]

Rule 5

The @. macro may be used with a symbol to broadcast that function over the preceding result.

@with begin
    [1, 2, 3]
    @. sqrt
end

is equivalent to

@with begin
    [1, 2, 3]
    sqrt.(_)
end

Gotchas for Julia users

Everything is a macro

Comma is used for options

Automatic variable name substitution

Automatic vectorization

Handling missing values

Gotchas for Stata users

All commands begin with @

@collapse has same syntax as @egen